refactor(racingCar): controller 분리
